workload/rand: refactor schema handling for reuse
The rand workload has logic for loading a table from a cluster
and adapting the randgen package for use with a gosql client. This
change extracts that logic into a reusable `LoadTable` function. The
utility will be used by LDR testing that generates high rates of
conflict for random schemas.

Release note: none
