+++ pkg/workload/rand/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "rand",
-    srcs = ["rand.go"],
+    srcs = [
+        "rand.go",
+        "schema.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/workload/rand",
     visibility = ["//visibility:public"],
     deps = [
